+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added a new class on the Python framework: `InputValidation`.
 - New request: Delete a list of groups - `DELETE /agents/groups - Params: {"ids":["id_1",...,"id_n"]}`
+- Input Validation at framework level of `group_id` parameter.
 
 ### Fixed
 - When adding a new agent, now it checks that its name is different than manager's name.
@@ -15,7 +16,6 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - Issue issue when deleting and restarting a list of agents
 - Issue with socket comunication in authd.
-
 
 ## [v2.1.0]
 ### Added